requirements.txt
python-dotenv
pyTelegramBotAPI
openai==1.3.9



